+++ busTrackerApi/src/main/kotlin/busTrackerApi/routing/lines/LinesClient.kt
@@ -2,13 +2,11 @@ package busTrackerApi.routing.lines
 
 import arrow.core.continuations.either
 import busTrackerApi.db.getItinerariesByItineraryCode
-import busTrackerApi.db.getItineraryByFullLineCode
 import busTrackerApi.db.getShapesByItineraryCode
 import busTrackerApi.exceptions.BusTrackerException.NotFound
 import busTrackerApi.extensions.getWrapped
 import busTrackerApi.routing.Response.ResponseFlowJson
 import busTrackerApi.routing.Response.ResponseJson
-import busTrackerApi.routing.lines.bus.getItinerariesResponse
 import busTrackerApi.utils.Pipeline
 import io.ktor.http.*
 import io.ktor.http.content.*
@@ -18,24 +16,6 @@ import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.map
 import simpleJson.asJson
 
-suspend fun Pipeline.getItineraries() = either {
-    val lineCode = call.parameters.getWrapped("lineCode").bind()
-    val direction = call.parameters.getWrapped("direction").bind().toIntOrNull()
-        ?: shift<Nothing>(busTrackerApi.exceptions.BusTrackerException.BadRequest())
-
-    val itineraries = getItineraryByFullLineCode(lineCode, direction)
-        ?: getItinerariesResponse(lineCode).bind()
-            .sortedBy { it.stops.count() }
-            .firstOrNull { it.direction == direction - 1 }
-        ?: shift<Nothing>(NotFound("Itinerary not found"))
-
-    val itinerariesOrdered =
-        itineraries.copy(stops = itineraries.stops.distinctBy { it.fullStopCode to it.order }.sortedBy { it.order })
-
-    val json = itinerariesOrdered.let(::buildItineraryJson).asJson()
-    call.caching = CachingOptions(CacheControl.MaxAge(maxAgeSeconds = 60 * 60))
-    ResponseJson(json, HttpStatusCode.OK)
-}
 
 suspend fun Pipeline.getItinerariesByItineraryCode() = either {
     val itineraryCode = call.parameters.getWrapped("itineraryCode").bind()