+++ busTrackerApi/src/main/kotlin/busTrackerApi/routing/lines/bus/BusLinesClient.kt
@@ -2,9 +2,7 @@ package busTrackerApi.routing.lines.bus
 
 import arrow.core.continuations.either
 import busTrackerApi.db.getItinerariesByItineraryCode
-import busTrackerApi.db.getItineraryByFullLineCode
 import busTrackerApi.db.getRoute
-import busTrackerApi.exceptions.BusTrackerException.BadRequest
 import busTrackerApi.exceptions.BusTrackerException.NotFound
 import busTrackerApi.extensions.getWrapped
 import busTrackerApi.routing.Response.ResponseJson
@@ -20,37 +18,6 @@ import io.ktor.http.content.*
 import io.ktor.server.application.*
 import io.ktor.server.plugins.cachingheaders.*
 import kotlinx.coroutines.flow.firstOrNull
-
-suspend fun Pipeline.getLocations() = either {
-    val lineCode = call.parameters.getWrapped("lineCode").bind()
-    val direction = call.parameters.getWrapped("direction").bind().toIntOrNull() ?: shift<Nothing>(BadRequest())
-    val stopCode = call.request.queryParameters["stopCode"]
-
-    val fullStopCode = if (stopCode != null) createStopCode(busCodMode, stopCode) else null
-    val codMode = getCodModeFromLineCode(lineCode)
-
-    val itinerary = getItineraryByFullLineCode(lineCode, direction) ?: shift<Nothing>(NotFound())
-
-    val route = getRoute(lineCode).getOrNull()
-    val simpleLineCode = route?.simpleLineCode ?: getSimpleLineCodeFromLineCode(lineCode)
-    val routeCodMode = route?.codMode ?: busCodMode
-
-    val locations = VehicleLocations(
-        locations = getLocationsResponse(
-            itinerary,
-            lineCode,
-            codMode,
-            fullStopCode
-        ).bind().vehiclesLocation.vehicleLocation,
-        codMode = routeCodMode.toInt(),
-        lineCode = simpleLineCode,
-    )
-
-    val json = buildVehicleLocationJson(locations)
-
-    call.caching = CachingOptions(CacheControl.MaxAge(maxAgeSeconds = 10))
-    ResponseJson(json, HttpStatusCode.OK)
-}
 
 suspend fun Pipeline.getLocationsByItineraryCode() = either {
     val itineraryCode = call.parameters.getWrapped("itineraryCode").bind()