+++ busTrackerApi/src/main/kotlin/busTrackerApi/routing/lines/bus/BusLinesController.kt
@@ -3,14 +3,11 @@ package busTrackerApi.routing.lines.bus
 import arrow.core.Either
 import busTrackerApi.config.EnvVariables.timeoutSeconds
 import busTrackerApi.db.models.ItineraryWithStops
-import busTrackerApi.db.models.StopOrder
-import busTrackerApi.exceptions.BusTrackerException
 import busTrackerApi.exceptions.BusTrackerException.SoapError
 import busTrackerApi.extensions.getSuspend
 import busTrackerApi.utils.auth
 import busTrackerApi.utils.defaultClient
 import busTrackerApi.utils.mapExceptionsF
-import crtm.soap.LineItineraryRequest
 import crtm.soap.LineLocationRequest
 import kotlinx.coroutines.withTimeoutOrNull
 
@@ -28,32 +25,3 @@ suspend fun getLocationsResponse(itinerary: ItineraryWithStops, lineCode: String
             getSuspend(lineRequest, defaultClient.value()::getLineLocationAsync)
         } ?: throw SoapError("Server error")
     }.mapLeft(mapExceptionsF)
-
-suspend fun getItinerariesResponse(lineCode: String) = Either.catch {
-    val itineraryRequest = LineItineraryRequest().apply {
-        codLine = lineCode
-        authentication = defaultClient.value().auth()
-        active = 1
-    }
-
-    withTimeoutOrNull(timeoutSeconds) {
-        getSuspend(itineraryRequest, defaultClient.value()::getLineItinerariesAsync)
-            .takeIf { it.itineraries.lineItinerary.isNotEmpty() }
-            ?: throw BusTrackerException.NotFound("No locations found for line $lineCode")
-    } ?: throw SoapError("Server error")
-
-}
-    .mapLeft(mapExceptionsF)
-    .map { itinerary ->
-        itinerary.itineraries.lineItinerary.map {
-            ItineraryWithStops(
-                itineraryCode = it.codItinerary,
-                fullLineCode = lineCode,
-                direction = it.direction - 1,
-                stops = it.stops.shortStop.mapIndexed { index, stop ->
-                    StopOrder(stop.codStop, null, index)
-                },
-                tripId = null
-            )
-        }
-    }