+++ busTrackerApi/src/main/kotlin/busTrackerApi/routing/lines/emt/EmtLinesClient.kt
@@ -3,10 +3,8 @@ package busTrackerApi.routing.lines.emt
 import arrow.core.Either
 import arrow.core.continuations.either
 import busTrackerApi.db.getItinerariesByItineraryCode
-import busTrackerApi.db.getItineraryByFullLineCode
 import busTrackerApi.db.getRoute
 import busTrackerApi.exceptions.BusTrackerException
-import busTrackerApi.exceptions.BusTrackerException.BadRequest
 import busTrackerApi.exceptions.BusTrackerException.NotFound
 import busTrackerApi.extensions.getWrapped
 import busTrackerApi.routing.Response
@@ -22,33 +20,6 @@ import io.ktor.http.content.*
 import io.ktor.server.application.*
 import io.ktor.server.plugins.cachingheaders.*
 import kotlinx.coroutines.flow.firstOrNull
-
-suspend fun Pipeline.getLocations(): Either<BusTrackerException, Response> = either {
-    val lineCode = call.parameters.getWrapped("lineCode").bind()
-    val direction =
-        call.parameters.getWrapped("direction").bind().toIntOrNull() ?: shift<Nothing>(BadRequest())
-    val stopCode = call.request.queryParameters["stopCode"]
-
-    val fullStopCode = if (stopCode != null) createStopCode(emtCodMode, stopCode) else run {
-        val itinerary = getItineraryByFullLineCode(lineCode, direction) ?: shift<Nothing>(NotFound())
-        itinerary.stops[itinerary.stops.size - 2].fullStopCode
-    }
-    val route = getRoute(lineCode).getOrNull()
-    val simpleLineCode = route?.simpleLineCode ?: getSimpleLineCodeFromLineCode(lineCode)
-    val codMode = route?.codMode ?: emtCodMode
-
-    val locations = VehicleLocations(
-        locations = getLocationsResponse(fullStopCode, simpleLineCode).bind()
-            .filter { it.direction == direction },
-        codMode = codMode.toInt(),
-        lineCode = simpleLineCode
-    )
-
-    val json = buildVehicleLocationJson(locations)
-
-    call.caching = CachingOptions(CacheControl.MaxAge(maxAgeSeconds = 15))
-    ResponseJson(json, HttpStatusCode.OK)
-}
 
 suspend fun Pipeline.getLocationsByItineraryCode(): Either<BusTrackerException, Response> = either {
     val itineraryCode = call.parameters.getWrapped("itineraryCode").bind()